Move ClaimRedeemFlowStatus type to shared module
The type lived in claimRedeemProgressDrawer, which is lazily
imported. Importing it from there made Vite warn about pulling a
value out of a dynamically-imported chunk. Move it to types.ts and
update the importers to fix the warning.
