+++ web/src/components/swapForm/types.ts
@@ -1,5 +1,12 @@
 import type { TokenWithGateway } from "types";
 
+export type ClaimRedeemFlowStatus =
+  | "idle"
+  | "redeem-error"
+  | "redeem-ready"
+  | "redeemed"
+  | "redeeming";
+
 export type SwapFormState = {
   approve10x: boolean;
   fromInputValue: string;
